fix: Product 모듈 분리
