+++ kitchenpos-common/src/main/java/kitchenpos/common/domain/Name.java
