+++ kitchenpos-common/src/test/java/kitchenpos/common/PriceTest.java
