+++ kitchenpos-product/build.gradle
@@ -1,0 +1,7 @@
+bootJar {
+    enabled = false
+}
+
+jar {
+    enabled = true
+}
