+++ settings.gradle
@@ -1,1 +1,8 @@
 rootProject.name = 'kitchenpos'
+
+include "kitchenpos-common",
+        "kitchenpos-order",
+        "kitchenpos-menu",
+        "kitchenpos-product",
+        "kitchenpos-table"
+
